refactor(api): migrate instance audit plan endpoints to v2
- Move instance audit plan related endpoints from v1 to v2
- Update data structures and API responses for v2
- Deprecate v1 endpoints for instance audit plans
- Improve error handling and data validation in v2 endpoints
